Added comments explainin some code , implemented 80% of the home screen

+++ app/src/main/java/com/example/chatbot/main/data/database_messages/model/ThreadMetadata.kt
@@ -23,6 +23,6 @@ import androidx.room.PrimaryKey
 data class ThreadMetadata(
     @PrimaryKey val uid: Long,
     val sessionUid: Long,
-    val questionUid: Long,
+    val questionUid: Int,
     val instructionUid: Long
 )
